Buddha talk/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
